обновление nextgen

+++ wp-content/plugins/nextgen-gallery/products/photocrati_nextgen/modules/ngglegacy/admin/manage.php
@@ -17,7 +17,7 @@ class nggManageGallery {
 		// GET variables
 		if( isset($_GET['gid']) ) {
 			$this->gid  = (int) $_GET['gid'];
-			$this->gallery = C_Gallery_Mapper::get_instance()->find($this->gid);
+			$this->gallery = C_Gallery_Mapper::get_instance()->find($this->gid, TRUE);
 		}
 		if( isset($_GET['pid']) )
 			$this->pid  = (int) $_GET['pid'];
@@ -194,7 +194,7 @@ class nggManageGallery {
 
 		$output = array();
 
-		$output[] = "<a href='{$image_url}' class='shutter' title='{$filename}'>";
+		$output[] = "<a href='{$image_url}' class='thickbox' title='{$filename}'>";
 		$output[] = "<img class='thumb' src='{$thumb_url}' id='thumb{$picture->pid}'/>";
 		$output[] = "</a>";
 
@@ -302,7 +302,7 @@ class nggManageGallery {
 		$alt_text	= empty($picture->alttext) ? $picture->filename: $picture->alttext;
 		$title		= esc_attr(__('View', 'nggallery'). " \"{$alt_text}\"");
 
-		return "<a href='{$image_url}' class='shutter' title='{$title}'>{$label}</a>";
+		return "<a href='{$image_url}' class='thickbox' title='{$title}'>{$label}</a>";
 	}
 
 	function render_meta_action_link($id, $picture)
@@ -345,28 +345,43 @@ class nggManageGallery {
 
 	function render_recover_action_link($id, $picture)
 	{
-		if ( !file_exists( $picture->imagePath . '_backup' )) return FALSE;
+		if (!file_exists($picture->imagePath . '_backup'))
+            return FALSE;
 
-		$url		= wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=recoverpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_recoverpicture');
-		$title		= esc_attr__('Recover image from backup', 'nggallery');
-		$label		= esc_html__('Recover', 'nggallery');
-		$alttext	= empty($picture->alttext) ? $picture->filename : $picture->alttext;
-		$confirm	= addslashes(__("Recover", 'nggallery'). " \"{$alttext}\"?");
-		$onclick	= "javascript:if(!confirm(\"{$confirm}\")) return false";
+		$url      = wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=recoverpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_recoverpicture');
+		$title    = esc_attr__('Recover image from backup', 'nggallery');
+		$label    = esc_html__('Recover', 'nggallery');
+        $question = __('Recover', 'nggallery');
 
-		return "<a href='{$url}' onclick='{$onclick}' class='confirmrecover' title='{$title}'>{$label}</a>";
+		$alttext = empty($picture->alttext) ? $picture->filename : $picture->alttext;
+		$alttext = M_NextGen_Data::strip_html(html_entity_decode($alttext), TRUE);
+		$alttext = htmlentities($alttext, ENT_QUOTES|ENT_HTML401);
+
+        // Event handler is found in nextgen_admin_page.js
+		return "<a href='{$url}'
+                   class='confirmrecover'
+                   data-question='{$question}'
+                   data-text='{$alttext}'
+                   title='{$title}'>{$label}</a>";
 	}
 
 	function render_delete_action_link($id, $picture)
 	{
-		$url		= wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=delpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_delpicture');
-		$title		= esc_attr__('Delete image', 'nggallery');
-		$label		= esc_html__('Delete', 'nggallery');
-		$alttext	= empty($picture->alttext) ? $picture->filename : $picture->alttext;
-		$confirm	= addslashes(__("Delete", 'nggallery'). " \"{$alttext}\"?");
-		$onclick	= "javascript:if(!confirm(\"{$confirm}\")) return false;";
+		$url      = wp_nonce_url("admin.php?page=nggallery-manage-gallery&amp;mode=delpic&amp;gid={$picture->galleryid}&amp;pid={$picture->pid}", 'ngg_delpicture');
+		$title    = esc_attr__('Delete image', 'nggallery');
+		$label    = esc_html__('Delete', 'nggallery');
+		$question = __('Delete', 'nggallery');
 
-		return "<a href='{$url}' onclick='{$onclick}' class='submitdelete delete' title='{$title}'>{$label}</a>";
+		$alttext = empty($picture->alttext) ? $picture->filename : $picture->alttext;
+		$alttext = M_NextGen_Data::strip_html(html_entity_decode($alttext), TRUE);
+		$alttext = htmlentities($alttext, ENT_QUOTES|ENT_HTML401);
+
+		// Event handler is found in nextgen_admin_page.js
+		return "<a href='{$url}'
+                   class='submitdelete delete'
+                   data-question='{$question}'
+                   data-text='{$alttext}'
+                   title='{$title}'>{$label}</a>";
 	}
 
 	function render_image_row_header()
@@ -796,12 +811,15 @@ class nggManageGallery {
 			}
 		}
 
-		if (isset ($_POST['updatepictures']) )  {
-		// Update pictures
+		if (isset($_POST['updatepictures']))
+		{
+            // Update pictures
+			$success = FALSE;
 
 			check_admin_referer('ngg_updategallery');
 
-			if ( nggGallery::current_user_can( 'NextGEN Edit gallery options' )  && !isset ($_GET['s']) ) {
+			if (nggGallery::current_user_can('NextGEN Edit gallery options') && !isset($_GET['s']))
+			{
                 $tags = array('<a>', '<abbr>', '<acronym>', '<address>', '<b>', '<base>', '<basefont>', '<big>', '<blockquote>', '<br>', '<br/>', '<caption>', '<center>', '<cite>', '<code>', '<col>', '<colgroup>', '<dd>', '<del>', '<dfn>', '<dir>', '<div>', '<dl>', '<dt>', '<em>', '<fieldset>', '<font>', '<h1>', '<h2>', '<h3>', '<h4>', '<h5>', '<h6>', '<hr>', '<i>', '<img>', '<ins>', '<label>', '<legend>', '<li>', '<menu>', '<noframes>', '<noscript>', '<ol>', '<optgroup>', '<option>', '<p>', '<pre>', '<q>', '<s>', '<samp>', '<select>', '<small>', '<span>', '<strike>', '<strong>', '<sub>', '<sup>', '<table>', '<tbody>', '<td>', '<tfoot>', '<th>', '<thead>', '<tr>', '<tt>', '<u>', '<ul>');
 				$fields = array('title', 'galdesc');
 				
@@ -813,29 +831,43 @@ class nggManageGallery {
 					$html = strip_tags($html, implode('', $tags));
 					$_POST[$field] = $html;
 				}
+
+				$mapper = C_Gallery_Mapper::get_instance();
 				
 				// Update the gallery
-				$mapper = C_Gallery_Mapper::get_instance();
-				if ($entity = $mapper->find($this->gid)) {
-					foreach ($_POST as $key => $value) {
-						$entity->$key = $value;
-					}
-					$mapper->save($entity);
+				if (!$this->gallery)
+				{
+					$this->gallery = $mapper->find($this->gid, TRUE);
 				}
 
-				if ($entity->path == '../' || FALSE !== strpos($entity->path, '/../'))
-					nggGallery::show_message(sprintf(__('One or more "../" in Gallery paths could be unsafe and NextGen Gallery will not delete this gallery automatically', 'nggallery'), $entity->{$entity->id_field}));
+				if ($this->gallery)
+				{
+					foreach ($_POST as $key => $value) {
+						$this->gallery->$key = $value;
+					}
+					$mapper->save($this->gallery);
 
-                wp_cache_delete($this->gid, 'ngg_gallery');
+					if ($this->gallery->is_invalid())
+					{
+						foreach ($this->gallery->get_errors() as $property => $errors) {
+							foreach ($errors as $error) {
+								nggGallery::show_error($error);
+							}
+						}
+					}
 
+					wp_cache_delete($this->gid, 'ngg_gallery');
+					$success = $this->gallery->is_valid();
+				}
 			}
 
-			$this->update_pictures();
-
-			//hook for other plugin to update the fields
-			do_action('ngg_update_gallery', $this->gid, $_POST);
-
-			nggGallery::show_message(__('Updated successfully', 'nggallery'));
+            $pictures_updated = $this->update_pictures();
+			if ($success || $pictures_updated >= 1)
+			{
+				// Hook for other plugin to update the fields
+				do_action('ngg_update_gallery', $this->gid, $_POST);
+				nggGallery::show_message(__('Updated successfully', 'nggallery'));
+			}
 		}
 
 		if (isset ($_POST['scanfolder']))  {
@@ -991,7 +1023,9 @@ class nggManageGallery {
             if (!empty($results))
             {
                 foreach ($results as $term_id) {
-                    wp_delete_term($term_id, 'ngg_tag');
+                    $term_id = apply_filters('ngg_pre_delete_unused_term_id', $term_id);
+                    if (!empty($term_id))
+                        wp_delete_term($term_id, 'ngg_tag');
                 }
             }
 		}
